Add approved_at fields to Agreement model.
