+++ app/models/report.rb
@@ -5,18 +5,7 @@ class Report < ActiveRecord::Base
   belongs_to :user
 
   include IdOrdered
-
-  def approved?
-    !approved_at.nil?
-  end
-
-  def mid_year_approved?
-    !mid_year_approved_at.nil?
-  end
-
-  def end_year_approved?
-    !end_year_approved_at.nil?
-  end
+  include ApprovalStages
 
   def approve!(stage, comment)
     stage_prefix = stage == :initial ? '' : "#{stage}_"
@@ -28,4 +17,5 @@ class Report < ActiveRecord::Base
       "#{stage_prefix}approved_snapshot_smart" => smart
     )
   end
+
 end
