+++ db/migrate/20150219143632_add_approved_at_to_agreements.rb
@@ -1,0 +1,7 @@
+class AddApprovedAtToAgreements < ActiveRecord::Migration
+  def change
+    add_column :agreements, :approved_at, :datetime
+    add_column :agreements, :mid_year_approved_at, :datetime
+    add_column :agreements, :end_year_approved_at, :datetime
+  end
+end
