+++ db/schema.rb
@@ -11,15 +11,18 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20150219140541) do
+ActiveRecord::Schema.define(version: 20150219143632) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
   create_table "agreements", force: :cascade do |t|
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at",           null: false
+    t.datetime "updated_at",           null: false
     t.integer  "user_id"
+    t.datetime "approved_at"
+    t.datetime "mid_year_approved_at"
+    t.datetime "end_year_approved_at"
   end
 
   add_index "agreements", ["user_id"], name: "index_agreements_on_user_id", using: :btree
