+++ spec/models/agreement_spec.rb
@@ -5,5 +5,6 @@ RSpec.describe Agreement, type: :model do
   it { is_expected.to belong_to(:user) }
 
   it_behaves_like 'id ordered'
+  include_examples 'approval'
 
 end
